requirements.txt
aiohttp==3.11.18
chumpy @ git+https://github.com/phangiaanh/W3Z_Packages.git@2fdf8b78afeae41d5c4b42f730c207c82e917cd2#subdirectory=chumpy
diffusers==0.32.2
einops==0.8.1
fvcore==0.1.5.post20221221
gdown==5.2.0
matplotlib==3.10.1
opencv-python==4.11.0.86
pip==25.0.1
plotly==6.0.1
PyQt6==6.8.1
PySocks==1.7.1
pytorch-lightning==2.5.1.post0
pytorch3d @ file:///content/pytorch3d-0.7.8-cp311-cp311-linux_x86_64.whl
tensorboard==2.19.0
torchaudio==2.6.0+cu118
torchgeometry @ git+https://github.com/phangiaanh/W3Z_Packages.git@f33ae7151913397b9e521f7fcac3b15c9196f877#subdirectory=torchgeometry
torchvision==0.21.0
trimesh==4.6.6
W3Z_AnimalPIPE @ git+https://github.com/phangiaanh/W3Z_AnimalPIPE.git@113df38c5ab7a2686cc69910db421d67d5dacd6d#egg=W3Z_AnimalPIPE
W3Z_AnimalSkeletons @ git+https://github.com/phangiaanh/W3Z_AnimalSkeletons.git@cd414bf30916c87301beffe5c95a482fc808af10#egg=W3Z_AnimalSkeletons
W3Z_AnimalTexture @ git+https://github.com/phangiaanh/W3Z_AnimalTexture.git@3d0ff4b0a2f45f07efc2f5781dbba80a1939b484#egg=W3Z_AnimalTexture
wheel==0.45.1
